+++ src/data/articles.ts
@@ -1,0 +1,11 @@
+export const articlesData = {
+  pytorch: [
+    { id: 1, title: "CIFAR10 Classifier", link: "#" },
+    { id: 2, title: "MNIST", link: "#" },
+    { id: 3, title: "RNN Part 1", link: "#" },
+    { id: 4, title: "RNN Part 2", link: "#" },
+    { id: 5, title: "RNN Part 3", link: "#" },
+    { id: 6, title: "Creating a Neural Network", link: "#" },
+  ],
+  kaggle: [{ id: 1, title: "Titanic Problem", link: "#" }],
+};
